chore: separate generateTableLabels to new file
